Category response maker
- New category response maker
- Removing transformers from patternMaker
- New category type article responders
- Event prices are now floats
- Globally using frameworks IOC container for object creation
- New category transformer to centralise transformation of category
objects
- Moved check for channel into responseMaker. Now throws exception if
no channel found
- New helper method for checking if a category belongs to a channel
- New IOC bindings
- Added new messages

+++ web/app/Api/Transformers/EventTransformer.php
@@ -44,7 +44,7 @@ class EventTransformer extends Transformer {
                 ,'showDate' => $event['show_date']
                 ,'showTime' => $event['show_time']
                 ,'epoch' => strtotime( $event['show_date'] . ' ' . $event['show_time'] )
-                ,'price' => number_format( $event['price'], 2 )
+                ,'price' => (float) number_format( $event['price'], 2 )
                 ,'url' => $event['url']
             ]
             ,'venue' => [
